Lots of small changes. Scaled everything down to integrate with JBox2D and physics is now working for box and circle colliders. Added some QoL keyboard shortcuts - Cmd+D to duplicate entities, backspace to delete entities. Gizmos are slightly buggy now. More to come.

+++ src/main/java/Components/GridLines.java
@@ -14,16 +14,16 @@ public class GridLines extends Component {
         Camera camera = Window.getScene().getCamera();
         Vector2f cameraPosition = camera.getPosition();
         Vector2f projectionSize = camera.getProjectionSize();
-        int firstX = ((int)(cameraPosition.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
-        int firstY = ((int)(cameraPosition.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
+        float firstX = ((int)(cameraPosition.x / Settings.GRID_WIDTH) - 1) * Settings.GRID_WIDTH;
+        float firstY = ((int)(cameraPosition.y / Settings.GRID_HEIGHT) - 1) * Settings.GRID_HEIGHT;
         int verticalLines = (int)(projectionSize.x * camera.getZoom() / Settings.GRID_WIDTH) + 2;
         int horizontalLines = (int)(projectionSize.y * camera.getZoom() / Settings.GRID_HEIGHT) + 2;
-        int height = (int)(projectionSize.y * camera.getZoom()) + Settings.GRID_WIDTH * 2;
-        int width = (int)(projectionSize.x * camera.getZoom()) + Settings.GRID_HEIGHT * 2;
+        float height = (int)(projectionSize.y * camera.getZoom()) + Settings.GRID_WIDTH * 2;
+        float width = (int)(projectionSize.x * camera.getZoom()) + Settings.GRID_HEIGHT * 2;
         int maxLines = Math.max(verticalLines, horizontalLines);
         for(int i = 0; i < maxLines; i++) {
-            int x = firstX + (Settings.GRID_WIDTH * i);
-            int y = firstY + (Settings.GRID_HEIGHT * i);
+            float x = firstX + (Settings.GRID_WIDTH * i);
+            float y = firstY + (Settings.GRID_HEIGHT * i);
             if(i < verticalLines)
                 DebugDraw.addLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), new Vector3f(0, 0, 0));
             if(i < horizontalLines)
